Extract the mock into its own submodule / subdirectory.
