+++ local-modules/codec/mocks/index.js
@@ -1,0 +1,9 @@
+// Copyright 2016-2017 the Bayou Authors (Dan Bornstein et alia).
+// Licensed AS IS and WITHOUT WARRANTY under the Apache License,
+// Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
+
+import MockCodable from './MockCodable';
+
+export {
+  MockCodable
+};
