+++ local-modules/codec/tests/test_Codec_decode.js
@@ -6,9 +6,9 @@ import { assert } from 'chai';
 import { before, describe, it } from 'mocha';
 
 import { Codec } from 'codec';
+import { MockCodable } from 'codec/mocks';
 import { FrozenBuffer } from 'util-common';
 
-import MockCodable from './MockCodable';
 
 describe('api-common/Codec.decode*()', () => {
   // Convenient bindings for `decode*()` and `encodeData()` to avoid a lot of
